introduce DockerExecutor and ComposeExecutor to simplify ComposeUp task
