+++ src/main/groovy/com/avast/gradle/dockercompose/tasks/ComposeUp.groovy
@@ -1,20 +1,10 @@
 package com.avast.gradle.dockercompose.tasks
 
-import com.avast.gradle.dockercompose.ComposeExtension
-import com.avast.gradle.dockercompose.NoOpLogger
-import com.avast.gradle.dockercompose.ServiceHost
-import com.avast.gradle.dockercompose.ServiceHostType
-import com.avast.gradle.dockercompose.ServiceInfo
-import com.avast.gradle.dockercompose.ContainerInfo
+import com.avast.gradle.dockercompose.*
 import org.gradle.api.DefaultTask
-import org.gradle.api.logging.Logger
 import org.gradle.api.tasks.TaskAction
-import org.gradle.process.ExecSpec
-import org.gradle.util.VersionNumber
-import org.yaml.snakeyaml.Yaml
 
 import java.time.Instant
-import java.util.concurrent.Executors
 
 class ComposeUp extends DefaultTask {
 
@@ -35,40 +25,30 @@ class ComposeUp extends DefaultTask {
     @TaskAction
     void up() {
         if (extension.buildBeforeUp) {
-            project.exec { ExecSpec e ->
-                extension.setExecSpecWorkingDirectory(e)
-                e.environment = extension.environment
-                String[] args = ['build']
-                args += extension.startedServices
-                e.commandLine extension.composeCommand(args)
-            }
+            extension.composeExecutor.execute('build', *extension.startedServices)
         }
-        project.exec { ExecSpec e ->
-            extension.setExecSpecWorkingDirectory(e)
-            e.environment = extension.environment
-            String[] args = ['up', '-d']
-            if (extension.removeOrphans()) {
-                args += '--remove-orphans'
-            }
-            if (extension.forceRecreate) {
-                args += '--force-recreate'
-            }
-            if (extension.scale()) {
-                args += extension.scale.collect { service, value ->
-                    ['--scale', "$service=$value"]
-                }.flatten()
-            }
-            args += extension.startedServices
-            e.commandLine extension.composeCommand(args)
+        String[] args = ['up', '-d']
+        if (extension.removeOrphans()) {
+            args += '--remove-orphans'
         }
+        if (extension.forceRecreate) {
+            args += '--force-recreate'
+        }
+        if (extension.scale()) {
+            args += extension.scale.collect { service, value ->
+                ['--scale', "$service=$value"]
+            }.flatten()
+        }
+        args += extension.startedServices
+        extension.composeExecutor.execute(args)
         try {
             if (extension.captureContainersOutput) {
-                captureContainersOutput(logger.&lifecycle)
+                extension.composeExecutor.captureContainersOutput(logger.&lifecycle)
             }
             if (extension.captureContainersOutputToFile != null) {
                 def logFile = extension.captureContainersOutputToFile
                 logFile.parentFile.mkdirs()
-                captureContainersOutput({ logFile.append(it + '\n') })
+                extension.composeExecutor.captureContainersOutput({ logFile.append(it + '\n') })
             }
             servicesInfos = loadServicesInfo().collectEntries { [(it.name): (it)] }
             waitForHealthyContainers(servicesInfos.values())
@@ -82,47 +62,12 @@ class ComposeUp extends DefaultTask {
         }
     }
 
-    protected void captureContainersOutput(Closure<Void> logMethod) {
-        // execute daemon thread that executes `docker-compose logs -f --no-color`
-        // the -f arguments means `follow` and so this command ends when docker-compose finishes
-        def t = Executors.defaultThreadFactory().newThread(new Runnable() {
-            @Override
-            void run() {
-                project.exec { ExecSpec e ->
-                    extension.setExecSpecWorkingDirectory(e)
-                    e.environment = extension.environment
-                    e.commandLine extension.composeCommand('logs', '-f', '--no-color')
-                    e.standardOutput = new OutputStream() {
-                        def buffer = new ArrayList<Byte>()
-
-                        @Override
-                        void write(int b) throws IOException {
-                            // store bytes into buffer until end-of-line character is detected
-                            if (b == 10 || b == 13) {
-                                if (buffer.size() > 0) {
-                                    // convert the byte buffer to characters and print these characters
-                                    def toPrint = buffer.collect { it as byte }.toArray() as byte[]
-                                    logMethod(new String(toPrint))
-                                    buffer.clear()
-                                }
-                            } else {
-                                buffer.add(b as Byte)
-                            }
-                        }
-                    }
-                }
-            }
-        })
-        t.daemon = true
-        t.start()
-    }
-
     protected Iterable<ServiceInfo> loadServicesInfo() {
-        getServiceNames().collect { createServiceInfo(it) }
+        extension.composeExecutor.getServiceNames().collect { createServiceInfo(it) }
     }
 
     protected ServiceInfo createServiceInfo(String serviceName) {
-        Iterable<String> containerIds = getContainerIds(serviceName)
+        Iterable<String> containerIds = extension.composeExecutor.getContainerIds(serviceName)
         Map<String, ContainerInfo> containerInfos = createContainerInfos(containerIds, serviceName)
         new ServiceInfo(name: serviceName, containerInfos: containerInfos)
     }
@@ -130,10 +75,10 @@ class ComposeUp extends DefaultTask {
     Map<String, ContainerInfo> createContainerInfos(Iterable<String> containerIds, String serviceName) {
         containerIds.collectEntries { String containerId ->
             logger.info("Container ID of service $serviceName is $containerId")
-            def inspection = getValidDockerInspection(serviceName, containerId)
-            ServiceHost host = getServiceHost(serviceName, inspection)
+            def inspection = extension.dockerExecutor.getValidDockerInspection(serviceName, containerId)
+            ServiceHost host = extension.dockerExecutor.getContainerHost(inspection, serviceName, logger)
             logger.info("Will use $host as host of service $serviceName")
-            def tcpPorts = getTcpPortsMapping(serviceName, inspection, host)
+            def tcpPorts = extension.dockerExecutor.getTcpPortsMapping(serviceName, inspection, host)
             String instanceName = inspection.Name.find(/${serviceName}_\d+/) ?: inspection.Name - '/'
             [(instanceName): new ContainerInfo(
                     instanceName: instanceName,
@@ -143,247 +88,12 @@ class ComposeUp extends DefaultTask {
         }
     }
 
-    Iterable<String> getServiceNames() {
-        if (extension.getDockerComposeVersion() >= VersionNumber.parse('1.6.0')) {
-            new ByteArrayOutputStream().withStream { os ->
-                project.exec { ExecSpec e ->
-                    extension.setExecSpecWorkingDirectory(e)
-                    e.environment = extension.environment
-                    e.commandLine extension.composeCommand('config', '--services')
-                    e.standardOutput = os
-                }
-                os.toString().readLines()
-            }
-        } else {
-            def composeFiles = extension.useComposeFiles.empty ? getStandardComposeFiles() : getCustomComposeFiles()
-            composeFiles.collectMany { composeFile ->
-                def compose = (Map<String, Object>) (new Yaml().load(project.file(composeFile).text))
-                // if there is 'version' on top-level then information about services is in 'services' sub-tree
-                compose.containsKey('version') ? ((Map) compose.get('services')).keySet() : compose.keySet()
-            }.unique()
-
-        }
-    }
-
-    Iterable<File> getStandardComposeFiles() {
-        def res = []
-        def f = findInParentDirectories('docker-compose.yml', project.projectDir)
-        if (f != null) res.add(f)
-        f = findInParentDirectories('docker-compose.override.yml', project.projectDir)
-        if (f != null) res.add(f)
-        res
-    }
-
-    Iterable<File> getCustomComposeFiles() {
-        extension.useComposeFiles.collect {
-            def f = project.file(it)
-            if (!f.exists()) {
-                throw new IllegalArgumentException("Custom Docker Compose file not found: $f")
-            }
-            f
-        }
-    }
-
-    File findInParentDirectories(String filename, File directory) {
-        if ((directory) == null) return null
-        def f = new File(directory, filename)
-        f.exists() ? f : findInParentDirectories(filename, directory.parentFile)
-    }
-
-    Iterable<String> getContainerIds(String serviceName) {
-        new ByteArrayOutputStream().withStream { os ->
-            project.exec { ExecSpec e ->
-                extension.setExecSpecWorkingDirectory(e)
-                e.environment = extension.environment
-                e.commandLine extension.composeCommand('ps', '-q', serviceName)
-                e.standardOutput = os
-            }
-            os.toString().readLines()
-        }
-    }
-
-    Map<String, Object> getDockerInspection(String containerId) {
-        new ByteArrayOutputStream().withStream { os ->
-            project.exec { ExecSpec e ->
-                e.environment = extension.environment
-                e.commandLine extension.dockerCommand('inspect', containerId)
-                e.standardOutput os
-            }
-            def inspectionAsString = os.toString()
-            logger.debug("Inspection for container $containerId: $inspectionAsString")
-            (new Yaml().load(inspectionAsString))[0] as Map<String, Object>
-        }
-    }
-
-    Map<String, Object> getValidDockerInspection(String serviceName, String containerId) {
-        def dockerInspection = getDockerInspection(containerId)
-        validateDockerInspection(serviceName, dockerInspection)
-        dockerInspection
-    }
-
-    void validateDockerInspection(String serviceName, Map<String, Object> inspection) {
-        ServiceHost serviceHost
-        try {
-            serviceHost = getServiceHost(serviceName, inspection, NoOpLogger.INSTANCE)
-        } catch (Exception e) {
-            throw new RuntimeException("Error when getting service host of service $serviceName: ${e.message}\n${inspection.toString()}", e)
-        }
-        if (serviceHost.type != ServiceHostType.Host) {
-            Map<String, Object> portsFromConfig = inspection.Config.ExposedPorts ?: [:]
-            Map<String, Object> portsFromNetwork = inspection.NetworkSettings.Ports
-            def missingPorts = portsFromConfig.keySet().findAll { !portsFromNetwork.containsKey(it) }
-            if (!missingPorts.empty) {
-                throw new RuntimeException("These ports of service $serviceName are declared as exposed but cannot be found in NetworkSettings: ${missingPorts.join(', ')}\\n${inspection.toString()}")
-            }
-        }
-    }
-
-    String getContainerPlatform(Map<String, Object> inspection) {
-        def platform = inspection.Platform as String
-        platform ?: getDockerPlatform()
-    }
-
-    List<String> getDockerInfo() {
-        new ByteArrayOutputStream().withStream { os ->
-            project.exec { ExecSpec e ->
-                e.environment = extension.environment
-                e.commandLine extension.dockerCommand('info')
-                e.standardOutput os
-            }
-            def asString = os.toString()
-            logger.debug("Docker info: $asString")
-            asString.readLines()
-        }
-    }
-
-    String getDockerPlatform() {
-        String osType = getDockerInfo().find { it.startsWith('OSType:') }
-        osType.empty ? System.getProperty("os.name") : osType.substring('OSType:'.length()).trim()
-    }
-
-    Map<String, Object> getNetworkInspection(String networkName) {
-        new ByteArrayOutputStream().withStream { os ->
-            project.exec { ExecSpec e ->
-                e.environment = extension.environment
-                e.commandLine extension.dockerCommand('network', 'inspect', networkName)
-                e.standardOutput os
-            }
-            def inspectionAsString = os.toString()
-            logger.debug("Inspection for network $networkName: $inspectionAsString")
-            (new Yaml().load(inspectionAsString))[0] as Map<String, Object>
-        }
-    }
-
-    String getNetworkGateway(String networkName) {
-        def networkInspection = getNetworkInspection(networkName)
-        if (networkInspection) {
-            Map<String, Object> ipam = networkInspection.IPAM
-            if (ipam) {
-                Map<String, Object>[] ipamConfig = ipam.Config
-                if (ipamConfig && ipamConfig.size() > 0) {
-                    return ipamConfig[0].Gateway
-                }
-            }
-        }
-        null
-    }
-
-    String getNetworkDriver(String networkName) {
-        def networkInspection = getNetworkInspection(networkName)
-        networkInspection ? networkInspection.Driver as String : ""
-    }
-
-    ServiceHost getServiceHost(String serviceName, Map<String, Object> inspection, Logger logger = this.logger) {
-        String servicesHost = extension.environment['SERVICES_HOST'] ?: System.getenv('SERVICES_HOST')
-        if (servicesHost) {
-            logger.lifecycle("SERVICES_HOST environment variable detected - will be used as hostname of service $serviceName ($servicesHost)'")
-            return new ServiceHost(host: servicesHost, type: ServiceHostType.RemoteDockerHost)
-        }
-        Map<String, Object> networkSettings = inspection.NetworkSettings
-        Map<String, Object> networks = networkSettings.Networks
-        Map.Entry<String, Object> firstNetworkPair = networks.find()
-        String dockerHost = extension.environment['DOCKER_HOST'] ?: System.getenv('DOCKER_HOST')
-        if (dockerHost) {
-            def host = dockerHost.toURI().host ?: 'localhost'
-            logger.lifecycle("DOCKER_HOST environment variable detected - will be used as hostname of service $serviceName ($host)'")
-            new ServiceHost(host: host, type: ServiceHostType.RemoteDockerHost)
-        } else if (isWindows() && getContainerPlatform(inspection).toLowerCase().contains('win') && "nat".equalsIgnoreCase(getNetworkDriver(firstNetworkPair.key))) {
-            logger.lifecycle("Will use direct access to the container of $serviceName")
-            return new ServiceHost(host: firstNetworkPair.value.IPAddress, type: ServiceHostType.DirectContainerAccess)
-        } else if (isMac() || isWindows()) {
-            logger.lifecycle("Will use localhost as host of $serviceName")
-            new ServiceHost(host: 'localhost', type: ServiceHostType.LocalHost)
-        } else {
-            // read gateway of first containers network
-            String gateway
-            if (networks && networks.every { it.key.toLowerCase().equals("host") }) {
-                gateway = 'localhost'
-                logger.lifecycle("Will use $gateway as host of $serviceName because it is using HOST network")
-                return new ServiceHost(host: 'localhost', type: ServiceHostType.Host)
-            } else if (networks && networks.size() > 0) {
-                gateway = firstNetworkPair.value.Gateway
-                if (!gateway) {
-                    logger.lifecycle("Gateway cannot be read from container inspection - trying to read from network inspection (network '${firstNetworkPair.key}')")
-                    gateway = getNetworkGateway(firstNetworkPair.key)
-                }
-                logger.lifecycle("Will use $gateway (network ${firstNetworkPair.key}) as host of $serviceName")
-            } else { // networks not specified (older Docker versions)
-                gateway = networkSettings.Gateway
-                logger.lifecycle("Will use $gateway as host of $serviceName")
-            }
-            if (!gateway) {
-                throw new RuntimeException('Gateway cannot be obtained')
-            }
-            new ServiceHost(host: gateway, type: ServiceHostType.NetworkGateway)
-        }
-    }
-
-    Map<Integer, Integer> getTcpPortsMapping(String serviceName, Map<String, Object> inspection, ServiceHost host) {
-        Map<Integer, Integer> ports = [:]
-        inspection.NetworkSettings.Ports.each { String exposedPortWithProtocol, forwardedPortsInfos ->
-            def (String exposedPortAsString, String protocol) = exposedPortWithProtocol.split('/')
-            if (!"tcp".equalsIgnoreCase(protocol)) {
-                return // from closure
-            }
-            int exposedPort = exposedPortAsString as int
-            if (!forwardedPortsInfos || forwardedPortsInfos.isEmpty()) {
-                logger.debug("No forwarded TCP port for service '$serviceName:$exposedPort'")
-            } else {
-                switch (host.type) {
-                    case ServiceHostType.LocalHost:
-                    case ServiceHostType.NetworkGateway:
-                    case ServiceHostType.RemoteDockerHost:
-                        if (forwardedPortsInfos.size() > 1) {
-                            logger.warn("More forwarded TCP ports for service '$serviceName:$exposedPort $forwardedPortsInfos'. Will use the first one.")
-                        }
-                        def forwardedPortInfo = forwardedPortsInfos.first()
-                        int forwardedPort = forwardedPortInfo.HostPort as int
-                        logger.info("Exposed TCP port on service '$serviceName:$exposedPort' will be available as $forwardedPort")
-                        ports.put(exposedPort, forwardedPort)
-                        break
-                    case ServiceHostType.Host:
-                        logger.info("Exposed TCP port on service '$serviceName:$exposedPort' will be available as $exposedPort because it uses HOST network")
-                        ports.put(exposedPort, exposedPort)
-                        break;
-                    case ServiceHostType.DirectContainerAccess:
-                        logger.info("Exposed TCP port on service '$serviceName:$exposedPort' will be available as $exposedPort because it uses direct access to the container")
-                        ports.put(exposedPort, exposedPort)
-                        break;
-                    default:
-                        throw new IllegalArgumentException("Unknown ServiceHostType '${host.type}' for service '$serviceName'")
-                        break
-                }
-            }
-        }
-        ports
-    }
-
     void waitForHealthyContainers(Iterable<ServiceInfo> servicesInfos) {
         def start = Instant.now()
         servicesInfos.forEach { serviceInfo ->
             serviceInfo.containerInfos.each { instanceName, containerInfo ->
                 while (true) {
-                    Map<String, Object> inspectionState = getDockerInspection(containerInfo.containerId).State
+                    Map<String, Object> inspectionState = extension.dockerExecutor.getInspection(containerInfo.containerId).State
                     if (inspectionState.containsKey('Health')) {
                         String healthStatus = inspectionState.Health.Status
                         if (!"starting".equalsIgnoreCase(healthStatus) && !"unhealthy".equalsIgnoreCase(healthStatus)) {
@@ -445,24 +155,5 @@ class ComposeUp extends DefaultTask {
                 }
             }
         }
-    }
-
-    String getServiceLogs(String containerId) {
-        new ByteArrayOutputStream().withStream { os ->
-            project.exec { ExecSpec e ->
-                e.environment = extension.environment
-                e.commandLine extension.dockerCommand('logs', '--follow=false', containerId)
-                e.standardOutput = os
-            }
-            os.toString().trim()
-        }
-    }
-
-    private static boolean isMac() {
-        System.getProperty("os.name").toLowerCase().startsWith("mac")
-    }
-
-    private static boolean isWindows() {
-        System.getProperty("os.name").toLowerCase().startsWith("win")
     }
 }
